requirements.txt
pymongo==4.2.0
requests==2.28.1
fake_useragent==0.1.11
scraperapi-sdk==0.2.2
beautifulsoup4==4.11.1